+++ photutils/psf/__init__.py
@@ -8,3 +8,5 @@ from .funcs import *
 from .groupstars import *
 from .matching import *
 from .models import *
+from .groupstars import *
+from .photometry import *