+++ CHANGES.rst
@@ -44,6 +44,9 @@ New Features
   - Added the ``param_uncert`` keyword to ``psf_photometry`` so that
     users can easily get uncertainties on fitted parameters. [#376]
 
+  - Added ``DAOPhotPSFPhotometry`` class to perform PSF photometry in
+    crowded fields. [#385]
+
 - ``photutils.psf_match``
 
   - Added ``create_matching_kernel`` and ``resize_psf`` functions.  Also